Extract logic from KtLint
