+++ ktlint-core/src/test/kotlin/com/pinterest/ktlint/core/ast/PackageKtTest.kt
@@ -12,7 +12,7 @@ import com.pinterest.ktlint.core.ast.ElementType.RPAR
 import com.pinterest.ktlint.core.ast.ElementType.VALUE_PARAMETER
 import com.pinterest.ktlint.core.ast.ElementType.VALUE_PARAMETER_LIST
 import com.pinterest.ktlint.core.ast.ElementType.WHITE_SPACE
-import com.pinterest.ktlint.core.internal.createRuleExecutionContext
+import com.pinterest.ktlint.core.internal.RuleExecutionContext.Companion.createRuleExecutionContext
 import org.assertj.core.api.Assertions.assertThat
 import org.jetbrains.kotlin.com.intellij.lang.ASTNode
 import org.jetbrains.kotlin.com.intellij.lang.FileASTNode
